requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
_pytorch_select=0.1=cpu_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
ca-certificates=2021.10.8=ha878542_0
certifi=2021.10.8=py38h578d9bd_1
charset-normalizer=2.0.9=pypi_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.28=pypi_0
debugpy=1.5.1=py38h295c915_0
decorator=4.4.2=pypi_0
entrypoints=0.3=pyhd8ed1ab_1003
fonttools=4.32.0=pypi_0
freetype=2.11.0=h70c0345_0
gensim=3.8.3=pypi_0
giflib=5.2.1=h7b6447c_0
googledrivedownloader=0.4=pypi_0
idna=3.3=pypi_0
imageio=2.13.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.6.0=py38he5a9106_0
ipython=7.30.1=py38h578d9bd_0
isodate=0.6.0=pypi_0
iteration-utilities=0.11.0=pypi_0
jedi=0.18.1=py38h578d9bd_0
jinja2=3.0.3=pypi_0
joblib=1.1.0=pypi_0
jpeg=9d=h7f8727e_0
jupyter_client=7.1.0=pyhd8ed1ab_0
jupyter_core=4.9.1=py38h578d9bd_1
kiwisolver=1.3.1=py38h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.0.1=pypi_0
matplotlib=3.5.0=pypi_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.4=pyhd8ed1ab_0
networkx=2.6.3=pypi_0
ninja=1.10.2=py38hd09550d_3
numba=0.54.1=pypi_0
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
olefile=0.46=pyhd3eb1b0_0
openssl=1.1.1n=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.3.4=pypi_0
parallel=20220222=ha770c72_0
parso=0.8.3=pyhd8ed1ab_0
perl=5.32.1=0_h7f98852_perl5
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.4.0=py38h5aabda8_0
pip=21.2.4=py38h06a4308_0
prompt-toolkit=3.0.24=pyha770c72_0
protobuf=4.0.0rc2=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pygments=2.10.0=pyhd8ed1ab_0
pyparsing=3.0.6=pypi_0
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.7.1=py3.8_cuda10.1.243_cudnn7.6.3_0
pytz=2021.3=pypi_0
pywavelets=1.2.0=pypi_0
pyyaml=6.0=pypi_0
pyzmq=19.0.2=py38ha71036d_2
rdflib=6.0.2=pypi_0
readline=8.1=h27cfd23_0
requests=2.26.0=pypi_0
scikit-image=0.19.0=pypi_0
scikit-learn=1.0.1=pypi_0
scipy=1.7.3=pypi_0
setuptools=58.0.4=py38h06a4308_0
setuptools-scm=6.4.2=pypi_0
six=1.16.0=pyhd3eb1b0_0
smart-open=5.2.1=pypi_0
sqlite=3.36.0=hc218d9a_0
tensorboardx=2.4.1=pypi_0
texttable=1.6.4=pypi_0
threadpoolctl=3.0.0=pypi_0
tifffile=2021.11.2=pypi_0
tk=8.6.11=h1ccaba5_0
tomli=2.0.1=pypi_0
torch-geometric=2.0.2=pypi_0
torch-geometric-temporal=0.51.0=pypi_0
torch-scatter=2.0.5=pypi_0
torch-sparse=0.6.12=pypi_0
torchaudio=0.7.2=py38
torchvision=0.8.2=cpu_py38ha229d99_0
tornado=6.1=py38h497a2fe_1
tqdm=4.62.3=pypi_0
traitlets=5.1.1=pyhd8ed1ab_0
typeguard=2.13.2=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.7=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_2
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pypi_0
zeromq=4.3.4=h9c3ff4c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
